WIP Prototype of running a flow with no project dependency
SimpleHubConfig is a new impl of HubConfig that does not have a hard dependency on a HubProject. It extends the new AbstractHubConfig class, to which I moved as much of HubConfigImpl as possible for reuse purposes. 

Modified a few clients of HubConfig that were casting to HubConfigImpl to call getMlUsername and getMlPassword - those are now part of the HubConfig interface so that those casts aren't needed. 

Added a new method to FlowRunnerImpl so that a Flow is retrieved from ML instead of the filesystem. 

Added StepDefinitionProvider so that StepRunnerFactory can get a StepDefinition from ML instead of from the filesystem. 

RunStepTest isn't yet a real test, it's just a sample program to show the new stuff in action
