+++ marklogic-data-hub/src/main/java/com/marklogic/hub/HubConfig.java
@@ -616,4 +616,8 @@ public interface HubConfig {
      * and then call this method.
      */
     void refreshProject();
+
+    String getMlUsername();
+
+    String getMlPassword();
 }
